implement array set

+++ src/vm/call_frame.rs
@@ -1,6 +1,6 @@
 #[derive(Clone)]
 pub struct CallFrame {
-    pub func_id: usize,
+    pub func_id: u64,
     pub instruction_index: usize,
     pub base_stack: usize,
 }
